jobCtrl

+++ srcs/shell/tc_signal.c
@@ -6,7 +6,7 @@
 /*   By: tgascoin <tgascoin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/21 15:45:36 by tgascoin          #+#    #+#             */
-/*   Updated: 2017/10/20 20:18:29 by kbagot           ###   ########.fr       */
+/*   Updated: 2017/10/24 20:05:29 by kbagot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,12 +23,39 @@ static void add_job_init(t_engine *e)
 	e->vm->job->idc = 1;
 }
 
-static void add_job(int npid, t_engine *e)
+static int	add_njob(t_job **tmp, t_engine *e, int npid)
 {
-	t_job	*tmp;
 	int		i;
 
 	i = 1;
+	*tmp = e->vm->job;
+	i++;
+	while ((*tmp)->next)
+	{
+		if ((*tmp)->id == npid)
+			return (1);
+		*tmp = (*tmp)->next;
+		i++;
+	}
+	if ((*tmp)->id)
+			return (1);
+	if (!((*tmp)->next = ft_memalloc(sizeof(t_job))))
+		return (1);
+	*tmp = (*tmp)->next;
+	(*tmp)->idc = i;
+	return (0);
+}
+
+static void print_job(t_job *tmp, int npid, t_engine *e)
+{
+	tmp->name = ft_strdup(e->buffer);
+	printf("\n[%d] %s  %s  %d\n", tmp->idc, "Stopped", tmp->name, npid);
+}
+
+static void add_job(int npid, t_engine *e)
+{
+	t_job	*tmp;
+
 	tmp = NULL;
 	if (!e->vm->job)
 	{
@@ -37,22 +64,15 @@ static void add_job(int npid, t_engine *e)
 	}
 	else if (e->vm->job)
 	{
-		tmp = e->vm->job;
-		i++;
-		while (tmp->next)
+		if (add_njob(&tmp, e, npid))
 		{
-			tmp = tmp->next;
-			i++;
-		}
-		if (!(tmp->next = ft_memalloc(sizeof(t_job))))
+			print_job(tmp, npid, e);
 			return ;
-		tmp = tmp->next;
-		tmp->idc = i;
+		}
 	}
-	tmp->name = ft_strdup(e->buffer);
+	print_job(tmp, npid, e);
 	tmp->id = npid;
 	tmp->next = NULL;
-	printf("\n[%d] %s  %s  %d\n", tmp->idc, "Stopped", tmp->name, npid);
 }
 
 void		tc_handle_signals(int sig)
@@ -69,7 +89,8 @@ void		tc_handle_signals(int sig)
 	else
 	{
 		g_last_signal = sig;
-		kill(g_pid, SIGSTOP);
+		if (g_pid != (int)getpid() && g_pid != 0)
+			kill(g_pid, SIGSTOP);
 	}
 }
 
@@ -78,10 +99,18 @@ void		tc_siglast(t_engine *engine)
 	(g_last_signal == SIGTSTP) ? add_job(g_pid, engine) : NULL;
 }
 
+void		tc_ign_exec()
+{
+	signal(SIGTSTP, SIG_IGN);
+	signal(SIGINT, SIG_IGN);
+	signal(SIGQUIT, SIG_IGN);
+}
+
 void		tc_stop_signals(void)
 {
 	g_last_signal = 0;
 	g_waitsig = 0;
+	//waitpid(pid, &status, WNOHANG); CHECK ALL JOBS STATUS
 	signal(SIGTSTP, &tc_handle_signals);
 }
